refactor(chat): share json-post headers via core/headers; drop chat→form coupling
Move xsrfToken from form-transport into core/headers and add jsonPostHeaders(accept)
so both the form and chat transports build the common CSRF/JSON/XHR header block from
one place. Re-export xsrfToken from form-transport for existing importers (action-form,
file-upload). The chat transport no longer imports anything from the form layer.
