+++ resources/js/core/headers.ts
@@ -11,3 +11,18 @@ export function withHeaders(
     ...headers,
   };
 }
+
+export function xsrfToken(): string {
+  const match = document.cookie.match(/XSRF-TOKEN=([^;]+)/);
+
+  return match ? decodeURIComponent(match[1]) : "";
+}
+
+export function jsonPostHeaders(accept: string): Record<string, string> {
+  return {
+    "Content-Type": "application/json",
+    Accept: accept,
+    "X-Requested-With": "XMLHttpRequest",
+    "X-XSRF-TOKEN": xsrfToken(),
+  };
+}
